Attribute Macro quest_hook::inline_hook::hook

#[hook]

Creates an inline hook at a C# method.

Panics

install will panic if the class or method was not found.

original will panic if the hook has not yet been installed.

Examples

use quest_hook::inline_hook::hook;
use quest_hook::libil2cpp::Il2CppObject;
use log::info;

#[hook("", "MainSettingsModelSO", "OnEnable")]
fn on_enable(this: &Il2CppObject) {
    info!("MainSettingsModelSO.OnEnable was called!");

    on_enable.original(this); // Call the original C# method
}

#[no_mangle]
pub extern "C" fn load() {
    info!("Installing hooks!");

    on_enable.install(); // Install the hook

    info!("Installed  hooks!");
}