Module quest_hook::libil2cpp::raw

Raw il2cpp types and functions

This module contains raw C types defined in libil2cpp and raw C functions dynamically loaded from libil2cpp.so.

Structs

CustomAttributesCache

EventInfo

FieldInfo

Il2CppAppContext

Il2CppAppDomain

Il2CppAppDomainSetup

Il2CppArgumentException

Il2CppArray

Il2CppArrayBounds

Il2CppArraySize

Il2CppArrayType

Il2CppAssembly

Il2CppAssemblyDefinition

Il2CppAssemblyName

Il2CppAssemblyNameDefinition

Il2CppAsyncCall

Il2CppAsyncResult

Il2CppCalendarData

Il2CppClass

Il2CppCodeGenOptions

Il2CppCodeRegistration

Il2CppCultureData

Il2CppCultureInfo

Il2CppCustomAttrInfo

Il2CppCustomAttributeTypeRange

Il2CppDateTimeFormatInfo

Il2CppDebuggerMetadataRegistration

Il2CppDecimal

Il2CppDecimal__bindgen_ty_1__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2__bindgen_ty_1

Il2CppDefaults

Il2CppDelegate

Il2CppDomain

Il2CppDouble

Il2CppErrorWrapper

Il2CppEventDefinition

Il2CppException

Il2CppExceptionWrapper

Il2CppFieldDefaultValue

Il2CppFieldDefinition

Il2CppFieldMarshaledSize

Il2CppFieldRef

Il2CppFileTime

Il2CppGenericClass

Il2CppGenericContainer

Il2CppGenericContext

Il2CppGenericInst

Il2CppGenericMethod

Il2CppGenericMethodFunctionsDefinitions

Il2CppGenericMethodIndices

Il2CppGenericParameter

Il2CppGlobalMetadataHeader

Il2CppGuid

Il2CppHStringHeader

Il2CppHString__

Il2CppIOAsyncResult

Il2CppIOSelectorJob

Il2CppIUnknown

Il2CppImage

Il2CppImageDefinition

Il2CppInterfaceOffsetPair

Il2CppInternalThread

Il2CppInteropData

Il2CppManagedMemorySnapshot

Il2CppManifestResourceInfo

Il2CppMarshalByRefObject

Il2CppMemoryCallbacks

Il2CppMetadataRegistration

Il2CppMetadataUsageList

Il2CppMetadataUsagePair

Il2CppMethodDefinition

Il2CppMethodExecutionContextInfo

Il2CppMethodExecutionContextInfoIndex

Il2CppMethodHeaderInfo

Il2CppMethodInfo

Il2CppMethodMessage

Il2CppMethodScope

Il2CppMethodSpec

Il2CppMulticastDelegate

Il2CppNameToTypeDefinitionIndexHashTable

Il2CppNumberFormatInfo

Il2CppObject

Il2CppParameterDefaultValue

Il2CppParameterDefinition

Il2CppPerfCounters

Il2CppProfiler

Il2CppPropertyDefinition

Il2CppPropertyInfo

Il2CppRGCTXDefinition

Il2CppRange

Il2CppReflectionAssembly

Il2CppReflectionAssemblyName

Il2CppReflectionEvent

Il2CppReflectionField

Il2CppReflectionGenericMethod

Il2CppReflectionMarshal

Il2CppReflectionMethod

Il2CppReflectionModule

Il2CppReflectionMonoEvent

Il2CppReflectionMonoEventInfo

Il2CppReflectionMonoType

Il2CppReflectionParameter

Il2CppReflectionPointer

Il2CppReflectionProperty

Il2CppReflectionRuntimeType

Il2CppReflectionType

Il2CppRegionInfo

Il2CppRuntimeInterfaceOffsetPair

Il2CppRuntimeStats

Il2CppSafeArray

Il2CppSafeArrayBound

Il2CppSafeHandle

Il2CppSequencePoint

Il2CppSequencePointIndex

Il2CppSequencePointSourceFile

Il2CppSingle

Il2CppSocketAddress

Il2CppSocketAsyncResult
Corresponds to Mono’s internal System.Net.Sockets.Socket.SocketAsyncResult class. Has no relation to Il2CppAsyncResult.

Il2CppSortKey

Il2CppStackFrame

Il2CppStackFrameInfo

Il2CppStatStg

Il2CppString

Il2CppStringBuilder

Il2CppStringLiteral

Il2CppSystemException

Il2CppThread

Il2CppType

Il2CppTypeDefinition

Il2CppTypeDefinitionMetadata

Il2CppTypeDefinitionSizes

Il2CppTypeSourceFilePair

Il2CppTypedRef

Il2CppVariant

Il2CppVariant__bindgen_ty_1___tagVARIANT

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1___tagBRECORD

Il2CppWaitHandle

Il2CppWin32Decimal

Il2CppWin32Decimal__bindgen_ty_1__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2__bindgen_ty_1

Il2CppWindowsRuntimeTypeName

Il2CppWindowsRuntimeTypeNamePair

MemberInfo

MethodInfo

MonitorData

ParameterInfo

PropertyInfo

VirtualInvokeData

_CRT_DOUBLE

_CRT_FLOAT

_LDBL12

_LDOUBLE

_LONGDOUBLE

_Mbstatet

__BindgenBitfieldUnit

__IncompleteArrayField

__crt_locale_data

__crt_locale_data_public

__crt_locale_pointers

__crt_multibyte_data

_div_t

_iobuf

_ldiv_t

_lldiv_t

max_align_t

Constants

ASSEMBLYREF_DISABLEJITCOMPILE_OPTIMIZER_FLAG

ASSEMBLYREF_ENABLEJITCOMPILE_TRACKING_FLAG

ASSEMBLYREF_FULL_PUBLIC_KEY_FLAG

ASSEMBLYREF_RETARGETABLE_FLAG

BUFSIZ

CHAR_BIT

CHAR_MAX

CHAR_MIN

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EAFNOSUPPORT

EAGAIN

EALREADY

EBADF

EBADMSG

EBUSY

ECANCELED

ECHILD

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EEXIST

EFAULT

EFBIG

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

ELOOP

EMFILE

EMLINK

EMSGSIZE

ENAMETOOLONG

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOBUFS

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENXIO

EOF

EOPNOTSUPP

EOTHER

EOVERFLOW

EOWNERDEAD

EPERM

EPIPE

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EROFS

ESPIPE

ESRCH

ETIME

ETIMEDOUT

ETXTBSY

EWOULDBLOCK

EXDEV

EXIT_FAILURE

EXIT_SUCCESS

FIELD_ATTRIBUTE_ASSEMBLY

FIELD_ATTRIBUTE_COMPILER_CONTROLLED

FIELD_ATTRIBUTE_FAMILY

FIELD_ATTRIBUTE_FAM_AND_ASSEM

FIELD_ATTRIBUTE_FAM_OR_ASSEM

FIELD_ATTRIBUTE_FIELD_ACCESS_MASK

FIELD_ATTRIBUTE_HAS_DEFAULT

FIELD_ATTRIBUTE_HAS_FIELD_MARSHAL

FIELD_ATTRIBUTE_HAS_FIELD_RVA

FIELD_ATTRIBUTE_INIT_ONLY

FIELD_ATTRIBUTE_LITERAL

FIELD_ATTRIBUTE_NOT_SERIALIZED

FIELD_ATTRIBUTE_PINVOKE_IMPL

FIELD_ATTRIBUTE_PRIVATE

FIELD_ATTRIBUTE_PUBLIC

FIELD_ATTRIBUTE_RESERVED_MASK

FIELD_ATTRIBUTE_RT_SPECIAL_NAME

FIELD_ATTRIBUTE_SPECIAL_NAME

FIELD_ATTRIBUTE_STATIC

FILENAME_MAX

FOPEN_MAX

IL2CPP_API_DYNAMIC_NO_DLSYM

IL2CPP_BIG_ENDIAN

IL2CPP_BINARY_SECTION_NAME

IL2CPP_BYTE_ORDER

IL2CPP_CAN_USE_MULTIPLE_SYMBOL_MAPS

IL2CPP_CXX_ABI_MSVC

IL2CPP_DEBUG

IL2CPP_DEVELOPMENT

IL2CPP_DIR_SEPARATOR

IL2CPP_DISABLE_FULL_MESSAGES

IL2CPP_ENABLE_DEFERRED_GC

IL2CPP_ENABLE_MONO_BUG_EMULATION

IL2CPP_ENABLE_PROFILER

IL2CPP_ENABLE_STACKTRACES

IL2CPP_EXCEPTION_DISABLED

IL2CPP_GC_BOEHM

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_CONTRAVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_COVARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_DEFAULT_CONSTRUCTOR_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NON_VARIANT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_NOT_NULLABLE_VALUE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_REFERENCE_TYPE_CONSTRAINT

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_SPECIAL_CONSTRAINT_MASK

IL2CPP_GENERIC_PARAMETER_ATTRIBUTE_VARIANCE_MASK

IL2CPP_HAS_DELETED_FUNCTIONS

IL2CPP_HAS_GC_DESCRIPTORS

IL2CPP_LITTLE_ENDIAN

IL2CPP_PAGE_SIZE

IL2CPP_PLATFORM_SUPPORTS_CPU_INFO

IL2CPP_PLATFORM_SUPPORTS_CUSTOM_SECTIONS

IL2CPP_PLATFORM_SUPPORTS_DEBUGGER_PRESENT

IL2CPP_PLATFORM_SUPPORTS_SYSTEM_CERTIFICATES

IL2CPP_PLATFORM_SUPPORTS_TIMEZONEINFO

IL2CPP_POINTER_SPARE_BITS

IL2CPP_SIZEOF_STRUCT_WITH_NO_INSTANCE_FIELDS

IL2CPP_SIZEOF_VOID_P

IL2CPP_TARGET_ANDROID

IL2CPP_TARGET_ARM64

IL2CPP_TARGET_ARMV7

IL2CPP_TARGET_DARWIN

IL2CPP_TARGET_IOS

IL2CPP_TARGET_JAVASCRIPT

IL2CPP_TARGET_LINUX

IL2CPP_TARGET_N3DS

IL2CPP_TARGET_OSX

IL2CPP_TARGET_PS4

IL2CPP_TARGET_PSP2

IL2CPP_TARGET_SWITCH

IL2CPP_TARGET_WINDOWS

IL2CPP_TARGET_WINDOWS_DESKTOP

IL2CPP_TARGET_WINRT

IL2CPP_TARGET_XBOXONE

IL2CPP_THREADS_STD

IL2CPP_THREAD_IMPL_HAS_COM_APARTMENTS

IL2CPP_USE_GENERIC_CPU_INFO

IL2CPP_USE_SEND_NOSIGNAL

IL2CPP_USE_STD_THREAD

IL2CPP_VALIDATE_FIELD_LAYOUT

IL2CPP_ZERO_LEN_ARRAY

INT_MAX

INT_MIN

Il2CppCallConvention_IL2CPP_CALL_C

Il2CppCallConvention_IL2CPP_CALL_DEFAULT

Il2CppCallConvention_IL2CPP_CALL_FASTCALL

Il2CppCallConvention_IL2CPP_CALL_STDCALL

Il2CppCallConvention_IL2CPP_CALL_THISCALL

Il2CppCallConvention_IL2CPP_CALL_VARARG

Il2CppCallType_Il2Cpp_CallType_BeginInvoke

Il2CppCallType_Il2Cpp_CallType_EndInvoke

Il2CppCallType_Il2Cpp_CallType_OneWay

Il2CppCallType_Il2Cpp_CallType_Sync

Il2CppCharSet_CHARSET_ANSI

Il2CppCharSet_CHARSET_NOT_SPECIFIED

Il2CppCharSet_CHARSET_UNICODE

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_EQ

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_GT

Il2CppDecimalCompareResult_IL2CPP_DECIMAL_CMP_LT

Il2CppGCEvent_IL2CPP_GC_EVENT_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_END

Il2CppGCEvent_IL2CPP_GC_EVENT_MARK_START

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_POST_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_START_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_PRE_STOP_WORLD

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_END

Il2CppGCEvent_IL2CPP_GC_EVENT_RECLAIM_START

Il2CppGCEvent_IL2CPP_GC_EVENT_START

Il2CppMetadataUsage_kIl2CppMetadataUsageFieldInfo

Il2CppMetadataUsage_kIl2CppMetadataUsageIl2CppType

Il2CppMetadataUsage_kIl2CppMetadataUsageInvalid

Il2CppMetadataUsage_kIl2CppMetadataUsageMethodDef

Il2CppMetadataUsage_kIl2CppMetadataUsageMethodRef

Il2CppMetadataUsage_kIl2CppMetadataUsageStringLiteral

Il2CppMetadataUsage_kIl2CppMetadataUsageTypeInfo

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_READ

Il2CppProfileFileIOKind_IL2CPP_PROFILE_FILEIO_WRITE

Il2CppProfileFlags_IL2CPP_PROFILE_ALLOCATIONS

Il2CppProfileFlags_IL2CPP_PROFILE_APPDOMAIN_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_ASSEMBLY_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_CLASS_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_ENTER_LEAVE

Il2CppProfileFlags_IL2CPP_PROFILE_EXCEPTIONS

Il2CppProfileFlags_IL2CPP_PROFILE_FILEIO

Il2CppProfileFlags_IL2CPP_PROFILE_GC

Il2CppProfileFlags_IL2CPP_PROFILE_GC_MOVES

Il2CppProfileFlags_IL2CPP_PROFILE_INLINING

Il2CppProfileFlags_IL2CPP_PROFILE_INS_COVERAGE

Il2CppProfileFlags_IL2CPP_PROFILE_IOMAP_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_JIT_COMPILATION

Il2CppProfileFlags_IL2CPP_PROFILE_METHOD_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MODULE_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_MONITOR_EVENTS

Il2CppProfileFlags_IL2CPP_PROFILE_NONE

Il2CppProfileFlags_IL2CPP_PROFILE_REMOTING

Il2CppProfileFlags_IL2CPP_PROFILE_STATISTICAL

Il2CppProfileFlags_IL2CPP_PROFILE_THREADS

Il2CppProfileFlags_IL2CPP_PROFILE_TRANSITIONS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_ARRAY

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_CLASS

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_INVALID

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_METHOD

Il2CppRGCTXDataType_IL2CPP_RGCTX_DATA_TYPE

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_ANOTHER_ASSEMBLY

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_EMBEDDED

Il2CppResourceLocation_IL2CPP_RESOURCE_LOCATION_IN_MANIFEST

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_CURRENT

Il2CppRuntimeUnhandledExceptionPolicy_IL2CPP_UNHANDLED_POLICY_LEGACY

Il2CppStat_IL2CPP_STAT_CLASS_STATIC_DATA_SIZE

Il2CppStat_IL2CPP_STAT_GENERIC_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_GENERIC_INSTANCE_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_INFLATED_TYPE_COUNT

Il2CppStat_IL2CPP_STAT_INITIALIZED_CLASS_COUNT

Il2CppStat_IL2CPP_STAT_METHOD_COUNT

Il2CppStat_IL2CPP_STAT_NEW_OBJECT_COUNT

Il2CppTypeEnum_IL2CPP_TYPE_ARRAY

Il2CppTypeEnum_IL2CPP_TYPE_BOOLEAN

Il2CppTypeEnum_IL2CPP_TYPE_BYREF

Il2CppTypeEnum_IL2CPP_TYPE_CHAR

Il2CppTypeEnum_IL2CPP_TYPE_CLASS

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_OPT

Il2CppTypeEnum_IL2CPP_TYPE_CMOD_REQD

Il2CppTypeEnum_IL2CPP_TYPE_END

Il2CppTypeEnum_IL2CPP_TYPE_ENUM

Il2CppTypeEnum_IL2CPP_TYPE_FNPTR

Il2CppTypeEnum_IL2CPP_TYPE_GENERICINST

Il2CppTypeEnum_IL2CPP_TYPE_I

Il2CppTypeEnum_IL2CPP_TYPE_I1

Il2CppTypeEnum_IL2CPP_TYPE_I2

Il2CppTypeEnum_IL2CPP_TYPE_I4

Il2CppTypeEnum_IL2CPP_TYPE_I8

Il2CppTypeEnum_IL2CPP_TYPE_INTERNAL

Il2CppTypeEnum_IL2CPP_TYPE_MODIFIER

Il2CppTypeEnum_IL2CPP_TYPE_MVAR

Il2CppTypeEnum_IL2CPP_TYPE_OBJECT

Il2CppTypeEnum_IL2CPP_TYPE_PINNED

Il2CppTypeEnum_IL2CPP_TYPE_PTR

Il2CppTypeEnum_IL2CPP_TYPE_R4

Il2CppTypeEnum_IL2CPP_TYPE_R8

Il2CppTypeEnum_IL2CPP_TYPE_SENTINEL

Il2CppTypeEnum_IL2CPP_TYPE_STRING

Il2CppTypeEnum_IL2CPP_TYPE_SZARRAY

Il2CppTypeEnum_IL2CPP_TYPE_TYPEDBYREF

Il2CppTypeEnum_IL2CPP_TYPE_U

Il2CppTypeEnum_IL2CPP_TYPE_U1

Il2CppTypeEnum_IL2CPP_TYPE_U2

Il2CppTypeEnum_IL2CPP_TYPE_U4

Il2CppTypeEnum_IL2CPP_TYPE_U8

Il2CppTypeEnum_IL2CPP_TYPE_VALUETYPE

Il2CppTypeEnum_IL2CPP_TYPE_VAR

Il2CppTypeEnum_IL2CPP_TYPE_VOID

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_ASSEMBLY_QUALIFIED

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_FULL_NAME

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_IL

Il2CppTypeNameFormat_IL2CPP_TYPE_NAME_FORMAT_REFLECTION

Il2CppVarType_IL2CPP_VT_ARRAY

Il2CppVarType_IL2CPP_VT_BLOB

Il2CppVarType_IL2CPP_VT_BLOB_OBJECT

Il2CppVarType_IL2CPP_VT_BOOL

Il2CppVarType_IL2CPP_VT_BSTR

Il2CppVarType_IL2CPP_VT_BSTR_BLOB

Il2CppVarType_IL2CPP_VT_BYREF

Il2CppVarType_IL2CPP_VT_CARRAY

Il2CppVarType_IL2CPP_VT_CF

Il2CppVarType_IL2CPP_VT_CLSID

Il2CppVarType_IL2CPP_VT_CY

Il2CppVarType_IL2CPP_VT_DATE

Il2CppVarType_IL2CPP_VT_DECIMAL

Il2CppVarType_IL2CPP_VT_DISPATCH

Il2CppVarType_IL2CPP_VT_EMPTY

Il2CppVarType_IL2CPP_VT_ERROR

Il2CppVarType_IL2CPP_VT_FILETIME

Il2CppVarType_IL2CPP_VT_HRESULT

Il2CppVarType_IL2CPP_VT_I1

Il2CppVarType_IL2CPP_VT_I2

Il2CppVarType_IL2CPP_VT_I4

Il2CppVarType_IL2CPP_VT_I8

Il2CppVarType_IL2CPP_VT_ILLEGAL

Il2CppVarType_IL2CPP_VT_ILLEGALMASKED

Il2CppVarType_IL2CPP_VT_INT

Il2CppVarType_IL2CPP_VT_INT_PTR

Il2CppVarType_IL2CPP_VT_LPSTR

Il2CppVarType_IL2CPP_VT_LPWSTR

Il2CppVarType_IL2CPP_VT_NULL

Il2CppVarType_IL2CPP_VT_PTR

Il2CppVarType_IL2CPP_VT_R4

Il2CppVarType_IL2CPP_VT_R8

Il2CppVarType_IL2CPP_VT_RECORD

Il2CppVarType_IL2CPP_VT_RESERVED

Il2CppVarType_IL2CPP_VT_SAFEARRAY

Il2CppVarType_IL2CPP_VT_STORAGE

Il2CppVarType_IL2CPP_VT_STORED_OBJECT

Il2CppVarType_IL2CPP_VT_STREAM

Il2CppVarType_IL2CPP_VT_STREAMED_OBJECT

Il2CppVarType_IL2CPP_VT_TYPEMASK

Il2CppVarType_IL2CPP_VT_UI1

Il2CppVarType_IL2CPP_VT_UI2

Il2CppVarType_IL2CPP_VT_UI4

Il2CppVarType_IL2CPP_VT_UI8

Il2CppVarType_IL2CPP_VT_UINT

Il2CppVarType_IL2CPP_VT_UINT_PTR

Il2CppVarType_IL2CPP_VT_UNKNOWN

Il2CppVarType_IL2CPP_VT_USERDEFINED

Il2CppVarType_IL2CPP_VT_VARIANT

Il2CppVarType_IL2CPP_VT_VECTOR

Il2CppVarType_IL2CPP_VT_VERSIONED_STREAM

Il2CppVarType_IL2CPP_VT_VOID

Il2CppWindowsRuntimeTypeKind_kTypeKindCustom

Il2CppWindowsRuntimeTypeKind_kTypeKindMetadata

Il2CppWindowsRuntimeTypeKind_kTypeKindPrimitive

LONG_MAX

LONG_MIN

L_tmpnam

L_tmpnam_s

MAXIMUM_NESTED_GENERICS_EXCEPTION_MESSAGE

MB_LEN_MAX

METHOD_ATTRIBUTE_ABSTRACT

METHOD_ATTRIBUTE_ASSEM

METHOD_ATTRIBUTE_COMPILER_CONTROLLED

METHOD_ATTRIBUTE_FAMILY

METHOD_ATTRIBUTE_FAM_AND_ASSEM

METHOD_ATTRIBUTE_FAM_OR_ASSEM

METHOD_ATTRIBUTE_FINAL

METHOD_ATTRIBUTE_HAS_SECURITY

METHOD_ATTRIBUTE_HIDE_BY_SIG

METHOD_ATTRIBUTE_MEMBER_ACCESS_MASK

METHOD_ATTRIBUTE_NEW_SLOT

METHOD_ATTRIBUTE_PINVOKE_IMPL

METHOD_ATTRIBUTE_PRIVATE

METHOD_ATTRIBUTE_PUBLIC

METHOD_ATTRIBUTE_REQUIRE_SEC_OBJECT

METHOD_ATTRIBUTE_RESERVED_MASK

METHOD_ATTRIBUTE_REUSE_SLOT

METHOD_ATTRIBUTE_RT_SPECIAL_NAME

METHOD_ATTRIBUTE_SPECIAL_NAME

METHOD_ATTRIBUTE_STATIC

METHOD_ATTRIBUTE_STRICT

METHOD_ATTRIBUTE_UNMANAGED_EXPORT

METHOD_ATTRIBUTE_VIRTUAL

METHOD_ATTRIBUTE_VTABLE_LAYOUT_MASK

METHOD_IMPL_ATTRIBUTE_CODE_TYPE_MASK

METHOD_IMPL_ATTRIBUTE_FORWARD_REF

METHOD_IMPL_ATTRIBUTE_IL

METHOD_IMPL_ATTRIBUTE_INTERNAL_CALL

METHOD_IMPL_ATTRIBUTE_MANAGED

METHOD_IMPL_ATTRIBUTE_MANAGED_MASK

METHOD_IMPL_ATTRIBUTE_MAX_METHOD_IMPL_VAL

METHOD_IMPL_ATTRIBUTE_NATIVE

METHOD_IMPL_ATTRIBUTE_NOINLINING

METHOD_IMPL_ATTRIBUTE_OPTIL

METHOD_IMPL_ATTRIBUTE_PRESERVE_SIG

METHOD_IMPL_ATTRIBUTE_RUNTIME

METHOD_IMPL_ATTRIBUTE_SYNCHRONIZED

METHOD_IMPL_ATTRIBUTE_UNMANAGED

MethodVariableKind_kMethodVariableKind_LocalVariable

MethodVariableKind_kMethodVariableKind_Parameter

MethodVariableKind_kMethodVariableKind_This

NEED_TO_ZERO_PTRFREE

NET_4_0

PARAM_ATTRIBUTE_HAS_DEFAULT

PARAM_ATTRIBUTE_HAS_FIELD_MARSHAL

PARAM_ATTRIBUTE_IN

PARAM_ATTRIBUTE_OPTIONAL

PARAM_ATTRIBUTE_OUT

PARAM_ATTRIBUTE_RESERVED_MASK

PARAM_ATTRIBUTE_UNUSED

PUBLIC_KEY_BYTE_LENGTH

RAND_MAX

SCHAR_MAX

SCHAR_MIN

SEEK_CUR

SEEK_END

SEEK_SET

SHRT_MAX

SHRT_MIN

STRUNCATE

SYS_OPEN

SequencePointKind_kSequencePointKind_Normal

SequencePointKind_kSequencePointKind_StepOut

THREAD_STATIC_FIELD_OFFSET

TMP_MAX

TMP_MAX_S

TYPE_ATTRIBUTE_ABSTRACT

TYPE_ATTRIBUTE_ANSI_CLASS

TYPE_ATTRIBUTE_AUTO_CLASS

TYPE_ATTRIBUTE_AUTO_LAYOUT

TYPE_ATTRIBUTE_BEFORE_FIELD_INIT

TYPE_ATTRIBUTE_CLASS

TYPE_ATTRIBUTE_CLASS_SEMANTIC_MASK

TYPE_ATTRIBUTE_EXPLICIT_LAYOUT

TYPE_ATTRIBUTE_FORWARDER

TYPE_ATTRIBUTE_HAS_SECURITY

TYPE_ATTRIBUTE_IMPORT

TYPE_ATTRIBUTE_INTERFACE

TYPE_ATTRIBUTE_LAYOUT_MASK

TYPE_ATTRIBUTE_NESTED_ASSEMBLY

TYPE_ATTRIBUTE_NESTED_FAMILY

TYPE_ATTRIBUTE_NESTED_FAM_AND_ASSEM

TYPE_ATTRIBUTE_NESTED_FAM_OR_ASSEM

TYPE_ATTRIBUTE_NESTED_PRIVATE

TYPE_ATTRIBUTE_NESTED_PUBLIC

TYPE_ATTRIBUTE_NOT_PUBLIC

TYPE_ATTRIBUTE_PUBLIC

TYPE_ATTRIBUTE_RESERVED_MASK

TYPE_ATTRIBUTE_RT_SPECIAL_NAME

TYPE_ATTRIBUTE_SEALED

TYPE_ATTRIBUTE_SEQUENTIAL_LAYOUT

TYPE_ATTRIBUTE_SERIALIZABLE

TYPE_ATTRIBUTE_SPECIAL_NAME

TYPE_ATTRIBUTE_STRING_FORMAT_MASK

TYPE_ATTRIBUTE_UNICODE_CLASS

TYPE_ATTRIBUTE_VISIBILITY_MASK

UCHAR_MAX

UINT_MAX

ULONG_MAX

USHRT_MAX

WCHAR_MAX

WCHAR_MIN

WINT_MAX

WINT_MIN

_ARGMAX

_ARM_WINAPI_PARTITION_DESKTOP_SDK_AVAILABLE

_CALL_REPORTFAULT

_CRT_BUILD_DESKTOP_APP

_CRT_FUNCTIONS_REQUIRED

_CRT_HAS_CXX17

_CRT_INTERNAL_NONSTDC_NAMES

_CRT_INTERNAL_PRINTF_LEGACY_MSVCRT_COMPATIBILITY

_CRT_INTERNAL_PRINTF_LEGACY_THREE_DIGIT_EXPONENTS

_CRT_INTERNAL_PRINTF_LEGACY_VSPRINTF_NULL_TERMINATION

_CRT_INTERNAL_PRINTF_LEGACY_WIDE_SPECIFIERS

_CRT_INTERNAL_PRINTF_STANDARD_SNPRINTF_BEHAVIOR

_CRT_INTERNAL_SCANF_LEGACY_MSVCRT_COMPATIBILITY

_CRT_INTERNAL_SCANF_LEGACY_WIDE_SPECIFIERS

_CRT_INTERNAL_SCANF_SECURECRT

_CRT_INTERNAL_STDIO_SYMBOL_PREFIX

_CRT_INT_MAX

_CRT_PACKING

_CRT_SECURE_CPP_OVERLOAD_SECURE_NAMES

_CRT_SECURE_CPP_OVERLOAD_SECURE_NAMES_MEMORY

_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES

_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES_COUNT

_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES_MEMORY

_CVTBUFSIZE

_HAS_CXX17

_HAS_CXX20

_HAS_EXCEPTIONS

_HAS_NODISCARD

_IOB_ENTRIES

_IOFBF

_IOLBF

_IONBF

_MAX_DIR

_MAX_DRIVE

_MAX_ENV

_MAX_EXT

_MAX_FNAME

_MAX_I64TOSTR_BASE2_COUNT

_MAX_I64TOSTR_BASE8_COUNT

_MAX_I64TOSTR_BASE10_COUNT

_MAX_I64TOSTR_BASE16_COUNT

_MAX_ITOSTR_BASE2_COUNT

_MAX_ITOSTR_BASE8_COUNT

_MAX_ITOSTR_BASE10_COUNT

_MAX_ITOSTR_BASE16_COUNT

_MAX_LTOSTR_BASE2_COUNT

_MAX_LTOSTR_BASE8_COUNT

_MAX_LTOSTR_BASE10_COUNT

_MAX_LTOSTR_BASE16_COUNT

_MAX_PATH

_MAX_U64TOSTR_BASE2_COUNT

_MAX_U64TOSTR_BASE8_COUNT

_MAX_U64TOSTR_BASE10_COUNT

_MAX_U64TOSTR_BASE16_COUNT

_MAX_ULTOSTR_BASE2_COUNT

_MAX_ULTOSTR_BASE8_COUNT

_MAX_ULTOSTR_BASE10_COUNT

_MAX_ULTOSTR_BASE16_COUNT

_NLSCMPERROR

_NSTREAM_

_OUT_TO_DEFAULT

_OUT_TO_MSGBOX

_OUT_TO_STDERR

_REPORT_ERRMODE

_SAL_VERSION

_SECURECRT_FILL_BUFFER_PATTERN

_SYS_OPEN

_TMP_MAX_S

_USE_ATTRIBUTES_FOR_SAL

_USE_DECLSPECS_FOR_SAL

_VCRT_COMPILER_PREPROCESSOR

_WRITE_ABORT_MSG

__GOT_SECURE_LIB__

__SAL_H_VERSION

__STDC_SECURE_LIB__

__STDC_WANT_SECURE_LIB__

ipv6AddressSize

kCustomAttributeIndexInvalid

kDefaultValueIndexNull

kEventIndexInvalid

kFieldIndexInvalid

kGenericContainerIndexInvalid

kGenericParameterIndexInvalid

kIl2CppInt32Max

kIl2CppInt32Min

kIl2CppInt64Max

kIl2CppInt64Min

kIl2CppIntPtrMax

kIl2CppIntPtrMin

kIl2CppOffsetOfArrayBounds

kIl2CppOffsetOfArrayLength

kIl2CppSizeOfArray

kIl2CppUInt32Max

kInteropDataIndexInvalid

kInvalidIl2CppMethodSlot

kMethodIndexInvalid

kPropertyIndexInvalid

kPublicKeyByteLength

kRGCTXIndexInvalid

kStringLiteralIndexInvalid

kTypeDefinitionIndexInvalid

kTypeIndexInvalid

Statics

__security_cookie

il2cpp_defaults

il2cpp_runtime_stats

kIl2CppNewLine

kIl2CppUInt64Max

kIl2CppUIntPtrMax

Traits

WrapRaw
Safe wrapper around a raw il2cpp type which can be used in its place

Functions

assembly_get_image

class_from_name

domain_get

domain_get_assemblies

Type Definitions

AssemblyIndex

CreateCCWFunc

CustomAttributeIndex

CustomAttributesCacheGenerator

DefaultValueDataIndex

DefaultValueIndex

EncodedMethodIndex

EventIndex

FILE

FieldIndex

GenericContainerIndex

GenericInstIndex

GenericMethodIndex

GenericParameterConstraintIndex

GenericParameterIndex

IL2CPP_VARIANT_BOOL

Il2CppCallConvention

Il2CppCallType

Il2CppChar

Il2CppCharSet

Il2CppDecimalCompareResult

Il2CppFrameWalkFunc

Il2CppGCEvent

Il2CppHString

Il2CppLogCallback

Il2CppMetadataUsage

Il2CppMethodPointer

Il2CppMethodSlot

Il2CppNativeChar

Il2CppProfileAllocFunc

Il2CppProfileFileIOFunc

Il2CppProfileFileIOKind

Il2CppProfileFlags

Il2CppProfileFunc

Il2CppProfileGCFunc

Il2CppProfileGCResizeFunc

Il2CppProfileMethodFunc

Il2CppProfileThreadFunc

Il2CppRGCTXDataType

Il2CppResourceLocation

Il2CppRuntimeUnhandledExceptionPolicy

Il2CppSetFindPlugInCallback

Il2CppStat

Il2CppTypeEnum

Il2CppTypeNameFormat

Il2CppVTable

Il2CppVarType

Il2CppWindowsRuntimeTypeKind

ImageIndex

InterfaceOffsetIndex

InterfacesIndex

InteropDataIndex

InvokerMethod

MethodIndex

MethodVariableKind

NestedTypeIndex

PInvokeMarshalCleanupFunc

PInvokeMarshalFromNativeFunc

PInvokeMarshalToNativeFunc

ParameterIndex

PropertyIndex

RGCTXIndex

SequencePointKind

StringIndex

StringLiteralIndex

TypeDefinitionIndex

TypeIndex

VTableIndex

_CoreCrtNonSecureSearchSortCompareFunction

_CoreCrtSecureSearchSortCompareFunction

__builtin_va_list

__crt_bool

__time32_t

__time64_t

__vcrt_bool

_invalid_parameter_handler

_locale_t

_onexit_t

_purecall_handler

div_t

errno_t

fpos_t

il2cpp_WorldChangedCallback

il2cpp_array_lower_bound_t

il2cpp_array_size_t

il2cpp_hresult_t

il2cpp_register_object_callback

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

ldiv_t

lldiv_t

mbstate_t

rsize_t

time_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

va_list

wchar_t

wctype_t

wint_t

Unions

Il2CppDecimal__bindgen_ty_1

Il2CppDecimal__bindgen_ty_2

Il2CppDouble_double

Il2CppHStringHeader__bindgen_ty_1

Il2CppObject__bindgen_ty_1

Il2CppRGCTXData

Il2CppRGCTXDefinitionData

Il2CppSingle_float

Il2CppType__bindgen_ty_1

Il2CppVariant__bindgen_ty_1

Il2CppVariant__bindgen_ty_1___tagVARIANT__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_1

Il2CppWin32Decimal__bindgen_ty_2

MethodInfo__bindgen_ty_1

MethodInfo__bindgen_ty_2