Trait quest_hook::libil2cpp::Arguments

pub unsafe trait Arguments<const N: usize> {
    type Type: Any;
    fn matches(args: &[&ParameterInfo]) -> bool;

    fn invokable(&self) -> [*mut c_void; N];
}

Trait implemented by types that can be used as a collection of C# method arguments

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the arguments, useful for caching

Loading content...
Required methods

fn matches(args: &[&ParameterInfo]) -> bool

Checks whether the type can be used as a C# argument collection with the given ParameterInfos to call a method

fn invokable(&self) -> [*mut c_void; N]

Returns an array of untyped pointer which can be used to invoke C# methods

Loading content...