Trait quest_hook::libil2cpp::CalleeArguments

pub unsafe trait CalleeArguments<const N: usize> {
    type Type: Any;
    fn matches(params: &[ParameterInfo]) -> bool;
}

Trait implemented by types that can be used as a collection of C# method parameters

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the parameters, useful for caching

Required methods

fn matches(params: &[ParameterInfo]) -> bool

Checks whether the type can be used as a C# parameter collection with the given ParameterInfos in a method definition