Trait quest_hook::libil2cpp::Return

pub unsafe trait Return {
    type Type: Any;
    fn matches(ty: &Il2CppType) -> bool;

    fn from_object(object: Option<&mut Il2CppObject>) -> Self;
}

Trait implemented by types that can be used as caller return types for C# methods

Safety

Interfaces depending on this trait assume that all of its methods are correct in an il2cpp context

Associated Types

type Type: Any

Normalized type of the return type, useful for caching

Loading content...
Required methods

fn matches(ty: &Il2CppType) -> bool

Checks whether the type can be used as a C# return type with the given Il2CppType in a method definition

fn from_object(object: Option<&mut Il2CppObject>) -> Self

Converts the Il2CppObject returned by runtime_invoke into self

Loading content...